Check for code_challenge before checking method
There was a bug when plain code challenge method is not supported based
on the configuration and PKCE is not essential, where all flows
authentication requests without PKCE would fail because plain is not
supported and is the default method.

It's fixed now, but maybe there is an underlying issue here, concerning
this use case; the PKCE RFC states that plain is the default method
and we follow it, however we provide the option to not support plain.
As a result each authentication request which has a code_challenge and
omits code_challenge_method will fail. Is that behaviour expected or we
should default in a code challenge method we support. Or should we
always support plain?